fix for APPLE
git-svn-id: svn://svn.savannah.nongnu.org/osip/trunk/osip@352 a617929b-4c5a-41d8-b177-2eda7db2ca6c

+++ src/osipparser2/osip_port.c
@@ -258,7 +258,7 @@ osip_strcasecmp (const char *s1, const char *s2)
 }
 
 int
-osip_strncasecmp (const char *s1, const char *s2, unsigned int len)
+osip_strncasecmp (const char *s1, const char *s2, size_t len)
 {
 #if (!defined WIN32 && !defined _WIN32_WCE)
   return strncasecmp (s1, s2, len);
